// Flance-main/apps/web/src/types/auth.ts
export type AppUser = {
  id: string;
  email: string;
  role: "CLIENT" | "FREELANCER" | "ADMIN";
  name?: string;
  avatarUrl?: string;
  headline?: string;
  services?: string;
  servicesTags?: string[];
  needs?: string;
  bio?: string;
  companyEnabled?: boolean;
  companyName?: string;
  companyCnpj?: string;
  companyDescription?: string;
  companyLocation?: string;
  companyCity?: string;
  companyState?: string;
  companyAddress?: string;
  companyWebsite?: string;
  companyInstagram?: string;
  companyWhatsapp?: string;
  companyEmail?: string;
  companyHours?: string;
  companyPhotos?: string[];
  companyIsOnline?: boolean;
  companyIsPhysical?: boolean;
  companyViews?: number;
  planTier?: string;
};